client: move FishingSpot to net.runelite.client.game
